Split out generated binding merge function.
