+++ src/utils/pause/mapScopes/index.js
@@ -4,12 +4,10 @@
 
 // @flow
 
-import { has } from "lodash";
 import {
   getScopes,
   type SourceScope,
-  type BindingData,
-  type BindingLocation
+  type BindingData
 } from "../../../workers/parser";
 import type { RenderableScope } from "../scopes/getScope";
 import { locColumn } from "./locColumn";
@@ -22,8 +20,10 @@ import {
   findGeneratedBindingForImportDeclaration,
   type GeneratedDescriptor
 } from "./findGeneratedBindingFromPosition";
-
-import { createObjectClient } from "../../../client/firefox";
+import {
+  buildGeneratedBindingList,
+  type GeneratedBindingLocation
+} from "./buildGeneratedBindingList";
 
 import { log } from "../../log";
 import type {
@@ -36,12 +36,6 @@ import type {
 } from "../../../types";
 
 export type OriginalScope = RenderableScope;
-
-export type GeneratedBindingLocation = {
-  name: string,
-  loc: BindingLocation,
-  desc: () => Promise<BindingContents | null>
-};
 
 function batchScopeMappings(
   originalAstScopes: Array<SourceScope>,
@@ -425,111 +419,4 @@ async function findGeneratedBinding(
     },
     expression: null
   };
-}
-
-function buildGeneratedBindingList(
-  scopes: Scope,
-  generatedAstScopes: SourceScope[],
-  thisBinding: ?BindingContents
-): Array<GeneratedBindingLocation> {
-  // The server's binding data doesn't include general 'this' binding
-  // information, so we manually inject the one 'this' binding we have into
-  // the normal binding data we are working with.
-  const frameThisOwner = generatedAstScopes.find(
-    generated => "this" in generated.bindings
-  );
-
-  let globalScope = null;
-  const clientScopes = [];
-  for (let s = scopes; s; s = s.parent) {
-    const bindings = s.bindings
-      ? Object.assign({}, ...s.bindings.arguments, s.bindings.variables)
-      : {};
-
-    clientScopes.push(bindings);
-    globalScope = s;
-  }
-
-  const generatedMainScopes = generatedAstScopes.slice(0, -2);
-  const generatedGlobalScopes = generatedAstScopes.slice(-2);
-
-  const clientMainScopes = clientScopes.slice(0, generatedMainScopes.length);
-  const clientGlobalScopes = clientScopes.slice(generatedMainScopes.length);
-
-  // Map the main parsed script body using the nesting hierarchy of the
-  // generated and client scopes.
-  const generatedBindings = generatedMainScopes.reduce((acc, generated, i) => {
-    const bindings = clientMainScopes[i];
-
-    if (generated === frameThisOwner && thisBinding) {
-      bindings.this = {
-        value: thisBinding
-      };
-    }
-
-    for (const name of Object.keys(generated.bindings)) {
-      // If there is no 'this' value, we exclude the binding entirely.
-      // Otherwise it would pass through as found, but "(unscoped)", causing
-      // the search logic to stop with a match.
-      if (name === "this" && !bindings[name]) {
-        continue;
-      }
-
-      const { refs } = generated.bindings[name];
-      for (const loc of refs) {
-        acc.push({
-          name,
-          loc,
-          desc: () => Promise.resolve(bindings[name] || null)
-        });
-      }
-    }
-    return acc;
-  }, []);
-
-  // Bindings in the global/lexical global of the generated code may or
-  // may not be the real global if the generated code is running inside
-  // of an evaled context. To handle this, we just look up the client scope
-  // hierarchy to find the closest binding with that name.
-  for (const generated of generatedGlobalScopes) {
-    for (const name of Object.keys(generated.bindings)) {
-      const { refs } = generated.bindings[name];
-      const bindings = clientGlobalScopes.find(b => has(b, name));
-
-      for (const loc of refs) {
-        if (bindings) {
-          generatedBindings.push({
-            name,
-            loc,
-            desc: () => Promise.resolve(bindings[name])
-          });
-        } else {
-          const globalGrip = globalScope && globalScope.object;
-          if (globalGrip) {
-            // Should always exist, just checking to keep Flow happy.
-
-            generatedBindings.push({
-              name,
-              loc,
-              desc: async () => {
-                const objectClient = createObjectClient(globalGrip);
-                return (await objectClient.getProperty(name)).descriptor;
-              }
-            });
-          }
-        }
-      }
-    }
-  }
-
-  // Sort so we can binary-search.
-  return generatedBindings.sort((a, b) => {
-    const aStart = a.loc.start;
-    const bStart = b.loc.start;
-
-    if (aStart.line === bStart.line) {
-      return locColumn(aStart) - locColumn(bStart);
-    }
-    return aStart.line - bStart.line;
-  });
 }
